Clarify content of message
Here you say that the message is BOTH the transaction component AND the salt commitment. But @nopara73 and I believe that the message is ONLY the transaction component, and the salt commitment is not blinded, but send in clear text to the server, alongside the minded message. This way the server knows the number of components for each salt commitment.